+++ internal/generator/constructor/doc.go
@@ -1,0 +1,2 @@
+// Package constructor generates a "constructor" function for each message type.
+package constructor