+++ internal/generator/generate.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 
 	"github.com/dave/jennifer/jen"
+	"github.com/dogmatiq/primo/internal/generator/constructor"
 	"github.com/dogmatiq/primo/internal/generator/exhaustiveswitch"
 	"github.com/dogmatiq/primo/internal/generator/internal/option"
 	"github.com/dogmatiq/primo/internal/generator/internal/scope"
@@ -34,6 +35,7 @@ func Generate(
 		ver,
 		req,
 		res,
+		constructor.Generate,
 		exhaustiveswitch.Generate,
 		mutator.Generate,
 	); err != nil {